Update README.md
1) Course Registration Bot:
Made this just as a fun project to see if I could skip waking up super early in the morning to register for projects. Used some basic Selenium and Python. Also the file secrets.json contains my username and password and has not been uploaded. You can do the same for your self